MemeMorph/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.0
numpy==1.25.2
opencv-python==4.8.0.76
transformers==4.32.1
torch==2.0.1
torchvision==0.15.2
scikit-image==0.21.0
pymongo==4.5.0
web3==6.8.0
python-jose==3.3.0